+++ core/src/rpcs/mina_archive.rs
@@ -20,7 +20,7 @@ impl MinaArchiveRPC {
     pub async fn detect_nori_burn(
         &self,
         contract_addr: &str,
-        from_height: u32,
+        from_height: u64,
     ) -> Result<Vec<BurnEvent>, Error> {
         detect_nori_burn(self.rpc_url.as_str(), contract_addr, from_height)
             .await
@@ -31,7 +31,7 @@ impl MinaArchiveRPC {
     /// node has no canonical block recorded at that height.
     pub async fn query_canonical_block_at_height(
         &self,
-        height: u32,
+        height: u64,
     ) -> Result<Option<String>, Error> {
         query_canonical_block_at_height(self.rpc_url.as_str(), height)
             .await
